refactor(space-sketch): extract the SVG canvas into <SpaceSketchCanvas>
Pull the ~150-line SVG render layer (grid, underlay, rooms at the chosen
boundary, leak diagnostics, and every live cue — hover, cut rubber-band, snap
rings, draw preview, delete telegraph, action-intent chip) out of the overlay
into a pure presentational SpaceSketchCanvas component with an explicit,
tsc-checked prop boundary. Shared UI types (Hover/SplitTarget/IntentTone) move
to space-sketch/types.ts; the colour language + room palette live with the
canvas. The overlay drops 1581 → 1287 lines and no longer mixes drawing with
interaction logic. No behavioural change. tsc clean; viewer suite 981/0/2.
